:pencil: change u32 to u64 (#391)
pull request #393

+++ src/mc_outputs.rs
@@ -54,7 +54,7 @@ pub fn read_max_spendable_mc_index(db: &Connection, kind: &str) -> Result<u32> {
 pub struct McIndexInterval {
     pub from_mci: u32,
     pub to_mci: u32,
-    pub accumulated: u32,
+    pub accumulated: i64,
     pub has_sufficient: bool,
 }
 
@@ -63,7 +63,7 @@ pub fn find_mc_index_interval_to_target_amount(
     kind: &str,
     address: &String,
     max_mci: u32,
-    target_amount: u32,
+    target_amount: u64,
 ) -> Result<Option<McIndexInterval>> {
     let from_mci = read_next_spendable_mc_index(db, kind, address, &[])?;
 
@@ -86,7 +86,7 @@ pub fn find_mc_index_interval_to_target_amount(
 
     //Original js has another implementation for mysql
     let min_mc_output = if kind == "witnessing" { 11.0 } else { 344.0 };
-    let max_count_outputs = (f64::from(target_amount) / min_mc_output).ceil() as u32;
+    let max_count_outputs = (target_amount as f64 / min_mc_output).ceil() as i64;
 
     let sql = format!(
         "SELECT main_chain_index, amount \
@@ -98,7 +98,7 @@ pub fn find_mc_index_interval_to_target_amount(
 
     struct Row {
         main_chain_index: u32,
-        amount: u32,
+        amount: i64,
     }
 
     let mut stmt = db.prepare_cached(&sql)?;
@@ -125,7 +125,7 @@ pub fn find_mc_index_interval_to_target_amount(
     for output in outputs {
         accumulated += output.amount;
         to_mci = output.main_chain_index;
-        if accumulated > target_amount {
+        if accumulated as u64 > target_amount {
             has_sufficient = true;
             break;
         }
